requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
passlib[bcrypt]
python-jose
redis
alembic
